Restructuring IRCCON
